+++ src/components/MemeCard/MemeCard.scss
