+++ src/components/PostNewMeme/PostNewMeme.scss
